device_service/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary
python-decouple==3.8
PyJWT==2.8.0
drf-spectacular==0.27.0
pika==1.3.2